+++ api_docs/changelog.md
@@ -20,6 +20,14 @@ format used by the Zulip server that they are interacting with.
 
 ## Changes in Zulip 10.0
 
+**Feature level 291**
+
+* `PATCH /realm`, [`GET /events`](/api/get-events),
+  [`POST /register`](/api/register-queue):
+  Added `can_delete_own_message_group` realm setting which is a
+  [group-setting value](/api/group-setting-values) describing the set of users
+  with permission to delete the messages that they have sent in the organization.
+
 **Feature level 290**
 
 * [`POST /user_groups/{user_group_id}/deactivate`](/api/deactivate-user-group):